move indexer into correct module
