+++ backend/event-listeners/src/bin/github-indexer/indexer/mod.rs
@@ -1,5 +1,6 @@
 pub mod composite;
 pub mod contributors;
+mod error;
 pub mod guarded;
 pub mod issues;
 pub mod logged;
@@ -10,10 +11,13 @@ pub mod with_state;
 
 use std::{
 	collections::hash_map::DefaultHasher,
+	fmt::Display,
 	hash::{Hash, Hasher},
 };
 
-use event_listeners::domain::{GithubEvent, IgnoreIndexerErrors, Indexer, IndexerResult as Result};
+use async_trait::async_trait;
+use error::Result;
+use event_listeners::domain::GithubEvent;
 
 pub trait Stateful<Id> {
 	fn store(&self, id: Id, events: &[GithubEvent]) -> anyhow::Result<()>;
@@ -24,3 +28,15 @@ pub fn hash<T: Hash>(t: &T) -> u64 {
 	t.hash(&mut s);
 	s.finish()
 }
+
+#[async_trait]
+pub trait Indexer<Id>: Send + Sync
+where
+	Id: Indexable,
+{
+	async fn index(&self, id: Id) -> Result<Vec<GithubEvent>>;
+}
+
+pub trait Indexable: Copy + Display + Send {}
+
+impl<I: Copy + Display + Send> Indexable for I {}
