+++ backend/event-listeners/src/bin/github-indexer/indexer/user.rs
@@ -7,7 +7,7 @@ use domain::{GithubFetchService, GithubFullUser, GithubUserId};
 use event_listeners::domain::{GithubEvent, GithubUserIndexRepository};
 use serde::{Deserialize, Serialize};
 
-use super::{hash, IgnoreIndexerErrors, Result};
+use super::{error::IgnoreErrors, hash, Result};
 
 #[derive(Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
 struct State {
