+++ backend/event-listeners/src/domain/mod.rs
@@ -10,12 +10,6 @@ pub use projectors::*;
 mod github;
 pub use github::Event as GithubEvent;
 
-mod indexer;
-pub use indexer::{
-	Error as IndexerError, IgnoreErrors as IgnoreIndexerErrors, Indexable, Indexer,
-	Result as IndexerResult,
-};
-
 mod repositories;
 pub use repositories::{
 	Error as RepositoryError, GithubRepoIndexRepository, GithubUserIndexRepository,
